requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.5
